start splitting input handler
