+++ crates/client/src/app/input_handler.rs
@@ -972,49 +972,8 @@ pub fn handle_input(app: &mut App, key: KeyCode) -> Result<(), Box<dyn Error>> {
         }
     }
 
-    // Handle hotbar assignment input when active
-    if let HotbarAssignmentState::ChoosingBlock {
-        hotbar_slot,
-        available_blocks,
-        selected_block,
-    } = &app.panels.hotbar_assignment
-    {
-        let slot = *hotbar_slot;
-        let blocks = available_blocks.clone();
-        let mut selected = *selected_block;
-
-        if app.ui.keybinds.matches("ui", "CLOSE_HELP_MENU", &key) {
-            app.panels.hotbar_assignment = HotbarAssignmentState::None;
-            app.core.game.res.log("Cancelled block assignment");
-            return Ok(());
-        }
-        if app.ui.keybinds.matches("movement", "MOVE_NORTH", &key) {
-            selected = selected.saturating_sub(1);
-        }
-        if app.ui.keybinds.matches("movement", "MOVE_SOUTH", &key) {
-            if selected < blocks.len().saturating_sub(1) {
-                selected += 1;
-            }
-        }
-        if app.ui.keybinds.matches("ui", "MENU_ACTIVATE", &key) || key == KeyCode::Enter {
-            if selected < blocks.len() {
-                let chosen_block = blocks[selected];
-                app.panels.build.hotbar_assignments[slot] = Some(chosen_block);
-                app.panels.hotbar_assignment = HotbarAssignmentState::None;
-                app.core.game.res.log(format!(
-                    "Assigned {} to slot F{}",
-                    lithicrivers_core::components::itemkind_name(chosen_block),
-                    slot + 1
-                ));
-            }
-            return Ok(());
-        }
-
-        app.panels.hotbar_assignment = HotbarAssignmentState::ChoosingBlock {
-            hotbar_slot: slot,
-            available_blocks: blocks,
-            selected_block: selected,
-        };
+    // Handle hotbar assignment input when active - delegated to input module
+    if crate::app::input::hotbar_assignment::handle_hotbar_assignment_input(app, key)? {
         return Ok(());
     }
