+++ crates/client/src/app/mod.rs
@@ -1,4 +1,5 @@
 pub mod app_initialization;
 pub mod input_handler;
+pub mod input;
 
 pub use input_handler::*;
